Factor out ready and htmlContentLoadComplete into LoadEvents to avoid race conditions with the global brackets property.
